Add trajectory to link E and G
