+++ vector_spaces/multiverse.py
@@ -2,31 +2,7 @@ import matplotlib.pyplot as plt
 import matplotlib as mpl
 import matplotlib.patheffects as pe
 import numpy as np
-
-
-def surface_1(x, y):
-    return np.abs(((1 / (1 + np.exp(1 - 12 * (x + 1))) * 1 / (1 + np.exp(1 - 3 * (y + 1))) * 1 / (1 + np.exp((x))) *
-                    (1 / (1 + np.exp(1 - 12 * ((y - 3) + 1))) * 1 / (1 + np.exp(1 - 3 * (y + 1))) * 1 / (
-                                1 + 2 * np.exp(((y - 3))))) * 2.5 +
-                    (1 / (1 + np.exp(1 - 12 * ((x - 2) + 1))) * 1 / (1 + np.exp(1 - (4 * y + 1))) * 1 / (
-                                1 + np.exp(((x - 2)))) / 2 +
-                     1 / (1 + np.exp(12 + 5 * x + 2 * y)) / 1.5)) - 1.1154647694013708e-07
-                   ) / (0.7427793855415158 - 1.1154647694013708e-07
-                        ))
-
-
-def gen_points(x1, x2, lpert1, upert1, lpert2, upert2, num):
-    for i in range(0, num):
-        pert1 = np.random.uniform(lpert1, upert1)
-        pert2 = np.random.uniform(lpert2, upert2)
-        p = surface_1(x2 + pert2, x1 + pert1)
-        if p > 1:
-            p = 1
-        if p < 0:
-            p = 0
-        colour = np.random.choice(['red', 'blue'], p=[p, 1 - p])
-        plt.plot([x1 + pert1], [x2 + pert2], marker="o", markersize=4, markeredgecolor='black', markerfacecolor=colour,
-                 alpha=0.25)
+import func.data_gen as func
 
 
 # Press the green button in the gutter to run the script.
@@ -38,7 +14,7 @@ if __name__ == '__main__':
     X, Y = np.meshgrid(x, y)
     for i in range(0, len(x)):
         for j in range(0, len(y)):
-            Z[i, j] = surface_1(X[0, i], Y[j, 0])
+            Z[i, j] = func.surface_1(X[0, i], Y[j, 0])
 
     origin = 'lower'
 
@@ -95,16 +71,23 @@ if __name__ == '__main__':
     plt.text(-1.45, -3.7, 'F', va='center', ha='center',
              rotation='horizontal', fontsize=font_size, color='white', alpha=.7,
              path_effects=[pe.withStroke(linewidth=4, foreground="black")])
-    plt.annotate("", xy=(-1.2, -2), xytext=(C1, C2),
+    plt.annotate("", xy=(-0.1, -1.2), xytext=(C1, C2),
+                 arrowprops=dict(arrowstyle="->", lw=2, color='yellow', alpha=0.7))
+    plt.annotate("", xy=(-1.2, -2), xytext=(-0.1, -1.2),
                  arrowprops=dict(arrowstyle="->", lw=2, color='yellow', alpha=0.7))
     plt.annotate("", xy=(-1.4, -3), xytext=(-1.2, -2),
                  arrowprops=dict(arrowstyle="->", lw=2, color='yellow', alpha=0.7))
     plt.annotate("", xy=(-1.6, -1.68), xytext=(-1.2, -2),
                  arrowprops=dict(arrowstyle="->", lw=2, color='yellow', alpha=0.7))
-    plt.text(-1.2-0.45, -1.75-0.7, r'E$_1$', va='center', ha='center',
+    plt.text(-1.2-0.45, -1.75-0.7, r'E$_2$', va='center', ha='center',
              rotation='horizontal', fontsize=font_size, color='yellow', alpha=.7,
              path_effects=[pe.withStroke(linewidth=4, foreground="black")])
-    plt.text(-1.2-0.6, -2 + 0.7, r'E$_2$', va='center', ha='center',
+    plt.text(-1.2-0.6, -2 + 0.7, r'E$_1$', va='center', ha='center',
+             rotation='horizontal', fontsize=font_size, color='yellow', alpha=.7,
+             path_effects=[pe.withStroke(linewidth=4, foreground="black")])
+    plt.annotate("", xy=(0, -1.5), xytext=(-0.1, -1.2),
+                 arrowprops=dict(arrowstyle="->", lw=2, color='yellow', alpha=0.7))
+    plt.text(0+0.4, -1.5-0.25, r'E$_3$', va='center', ha='center',
              rotation='horizontal', fontsize=font_size, color='yellow', alpha=.7,
              path_effects=[pe.withStroke(linewidth=4, foreground="black")])
     plt.annotate("", xy=(-1.8, -1.8), xytext=(C1, C2),
@@ -115,14 +98,15 @@ if __name__ == '__main__':
              rotation='horizontal', fontsize=font_size, color='white', alpha=.7,
              path_effects=[pe.withStroke(linewidth=4, foreground="black")])
 
-    gen_points(0, 0, -1.8, 1.8, -1.7, 1.7, 200)
-    gen_points(-3, 2, -0.9, 6, -1.5, 1.5, 100)
-    gen_points(-4, -4, 0, 3, 0, 1, 100)
-    gen_points(-1, -3, -0.5, 0.5, -0.5, 1.5, 30)
-    gen_points(-4, -2, 0, 2, -0.5, 1, 30)
-    gen_points(2.45, 1.35, -0.5, 0.5, -0.25, 0.25, 30)
-    gen_points(0.8 + 0.15, -3.4 + 1, -0.5, 0.5, -1.5, 1.5, 20)
-    gen_points(-0.5 + 0.25, -3 + 1, -0.5, 0.5, -1.5, 1.5, 20)
+    func.gen_points(0, 0, -1.8, 1.5, -1.7, 1.7, 200)
+    func.gen_points(-3, 2, -0.9, 6, -1, 1.5, 100)
+    func.gen_points(-4, -4, 0, 3, 0, 1, 100)
+    func.gen_points(-1, -3, -0.5, 0.5, -0.5, 1.5, 30)
+    func.gen_points(-4, -2, 0, 2, -0.5, 1, 30)
+    func.gen_points(2.45, 1.35, -1, 1, -0.25, 0.15, 50)
+    func.gen_points(0.8 + 0.15, -3.4 + 1, -0.5, 0.5, -1.5, 1.5, 20)
+    func.gen_points(-0.5 + 0.25, -3 + 1, -0.5, 0.5, -1.5, 1.5, 20)
+    func.gen_points(2.45 + 0.15, -0.25, -0.15, 0.25, -0.45, 0.35, 20)
     for i in range(0, 20):
         pert1 = np.random.uniform(-0.35, 0.35)
         pert2 = np.random.uniform(-0.35, 0.35)
